Add shop products
This commit adds shop products to the database and an appropriate endpoint to retrieve them. The products are managed via a yaml file in the data folder. On start of the backend the yaml file is synched to the database to update or delete the respective products. Later a on-demand command or time based event for the server can be introduced to keep the products up to date while the server is running.